COMP: Fix compilation on 64 bit systems. Create a better actor-address to selection-index map that handles 32 or 64 bit addresses cleanly.

+++ VTK/Rendering/vtkRenderer.cxx
@@ -1668,7 +1668,7 @@ int vtkRenderer::UpdateGeometryForSelection()
 
     if (this->SelectMode == vtkRenderer::COLOR_BY_ACTOR)
       {
-      this->IdentPainter->SetToColorByActorId((int)this->PropArray[i]);
+      this->IdentPainter->SetToColorByActorId(this->PropArray[i]);
       }
     else if (this->SelectMode == vtkRenderer::COLOR_BY_CELL_ID_HIGH ||
              this->SelectMode == vtkRenderer::COLOR_BY_CELL_ID_MID ||
